requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
pillow>=10.0.0
pytesseract>=0.3.10
flask>=2.3.0
werkzeug>=2.3.0
transformers>=4.30.0
torchvision>=0.15.0
openai>=1.0.0

